updated httpclient5 updated copyright refactor
