+++ cs-httpclient-commons/src/main/java/io/cloudslang/content/httpclient/entities/CustomConnectionManager.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2024 Open Text
+ * Copyright 2022-2025 Open Text
  * This program and the accompanying materials
  * are made available under the terms of the Apache License v2.0 which accompany this distribution.
  *
@@ -13,12 +13,11 @@
  * limitations under the License.
  */
 
-package io.cloudslang.content.httpclient.services;
+
+package io.cloudslang.content.httpclient.entities;
 
 import com.hp.oo.sdk.content.plugin.GlobalSessionObject;
 import com.hp.oo.sdk.content.plugin.SessionResource;
-import io.cloudslang.content.httpclient.entities.HttpClientInputs;
-import org.apache.hc.client5.http.HttpRoute;
 import org.apache.hc.client5.http.impl.io.PoolingHttpClientConnectionManager;
 import org.apache.hc.client5.http.socket.ConnectionSocketFactory;
 import org.apache.hc.client5.http.socket.PlainConnectionSocketFactory;
