+++ cs-httpclient-commons/src/main/java/io/cloudslang/content/httpclient/entities/CustomEntity.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2024 Open Text
+ * Copyright 2022-2025 Open Text
  * This program and the accompanying materials
  * are made available under the terms of the Apache License v2.0 which accompany this distribution.
  *
@@ -13,9 +13,9 @@
  * limitations under the License.
  */
 
-package io.cloudslang.content.httpclient.services;
 
-import io.cloudslang.content.httpclient.entities.HttpClientInputs;
+package io.cloudslang.content.httpclient.entities;
+
 import io.cloudslang.content.httpclient.utils.HttpUtils;
 import io.cloudslang.content.httpclient.utils.UrlEncodeException;
 import org.apache.commons.lang3.StringUtils;
