+++ cs-httpclient-commons/src/main/java/io/cloudslang/content/httpclient/services/HttpClientService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2024 Open Text
+ * Copyright 2022-2025 Open Text
  * This program and the accompanying materials
  * are made available under the terms of the Apache License v2.0 which accompany this distribution.
  *
@@ -14,11 +14,12 @@
  */
 
 
+
 package io.cloudslang.content.httpclient.services;
 
 import com.hp.oo.sdk.content.plugin.GlobalSessionObject;
 import com.hp.oo.sdk.content.plugin.SerializableSessionObject;
-import io.cloudslang.content.httpclient.entities.HttpClientInputs;
+import io.cloudslang.content.httpclient.entities.*;
 import io.cloudslang.content.httpclient.utils.ExecutionTimeout;
 import org.apache.hc.client5.http.auth.AuthCache;
 import org.apache.hc.client5.http.auth.CredentialsProvider;
